streak와 badge 관련 로직을 별도 파일로 추출
